# BackLegiaoTricolor/openapi.json
{"openapi":"3.0.1","info":{"title":"API Loja Tricolor","description":"Documentação gerada com Swagger e Springdoc OpenAPI","version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/users/{id}":{"get":{"tags":["user-controller"],"operationId":"findById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDTO"}}}}}},"put":{"tags":["user-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDTO"}}}}}},"delete":{"tags":["user-controller"],"operationId":"deactivate","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/orders/cart/items/{itemId}":{"put":{"tags":["order-controller"],"operationId":"updateItem","parameters":[{"name":"itemId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateItemRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderResponseDTO"}}}}}}},"/admin/products/{id}":{"put":{"tags":["admin-product-controller"],"operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponseDTO"}}}}}},"delete":{"tags":["admin-product-controller"],"operationId":"delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/webhooks/pagseguro":{"post":{"tags":["webhook-controller"],"operationId":"handlePagSeguroWebhook","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PagSeguroWebhookDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/users/client":{"post":{"tags":["user-controller"],"operationId":"createClient","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDTO"}}}}}}},"/users/admin":{"post":{"tags":["user-controller"],"operationId":"createAdmin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDTO"}}}}}}},"/payments/checkout":{"post":{"tags":["payment-controller"],"operationId":"checkout","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CheckoutRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CheckoutResponseDTO"}}}}}}},"/orders/cart/items":{"post":{"tags":["order-controller"],"operationId":"addItem","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddItemRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderResponseDTO"}}}}}}},"/orders/cart/checkout":{"post":{"tags":["order-controller"],"operationId":"checkout_1","responses":{"200":{"description":"OK"}}}},"/checkout":{"post":{"tags":["checkout-controller"],"operationId":"checkout_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CheckoutRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CheckoutResponseDTO"}}}}}}},"/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponseDTO"}}}}}}},"/auth/admin/login":{"post":{"tags":["auth-controller"],"operationId":"adminLogin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminLoginRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponseDTO"}}}}}}},"/admin/products":{"post":{"tags":["admin-product-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponseDTO"}}}}}}},"/admin/products/{productId}/images":{"get":{"tags":["product-image-controller"],"operationId":"list_1","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductImageDTO"}}}}}}},"post":{"tags":["product-image-controller"],"operationId":"addImage","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductImageRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductImageDTO"}}}}}}},"/orders/cart/shipping":{"patch":{"tags":["order-controller"],"operationId":"applyShipping","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderResponseDTO"}}}}}}},"/admin/products/{productId}/images/{imageId}/main":{"patch":{"tags":["product-image-controller"],"operationId":"setMainImage","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"imageId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/admin/products/{id}/stock":{"patch":{"tags":["admin-product-controller"],"operationId":"updateStock","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"quantity","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/admin/products/{id}/status":{"patch":{"tags":["admin-product-controller"],"operationId":"updateStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"active","in":"query","required":true,"schema":{"type":"boolean"}}],"responses":{"200":{"description":"OK"}}}},"/users":{"get":{"tags":["user-controller"],"operationId":"findAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponseDTO"}}}}}}}},"/products":{"get":{"tags":["public-product-controller"],"operationId":"list","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductResponseDTO"}}}}}}}},"/products/{id}":{"get":{"tags":["public-product-controller"],"operationId":"detail","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponseDTO"}}}}}}},"/orders/cart":{"get":{"tags":["order-controller"],"operationId":"getCart","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderResponseDTO"}}}}}}},"/admin/products/{productId}/images/{imageId}":{"delete":{"tags":["product-image-controller"],"operationId":"delete","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"imageId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"UserUpdateDTO":{"type":"object","properties":{"name":{"maxLength":100,"minLength":3,"type":"string"},"password":{"maxLength":2147483647,"minLength":6,"type":"string"}}},"UserResponseDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"cpf":{"type":"string"},"telefone":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["ADMIN","CLIENT"]}}},"UpdateItemRequestDTO":{"required":["quantity"],"type":"object","properties":{"quantity":{"type":"integer","format":"int32"}}},"OrderItemResponseDTO":{"type":"object","properties":{"itemId":{"type":"string","format":"uuid"},"productId":{"type":"string","format":"uuid"},"productName":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"price":{"type":"number"},"subtotal":{"type":"number"}}},"OrderResponseDTO":{"type":"object","properties":{"orderId":{"type":"string","format":"uuid"},"status":{"type":"string","enum":["CREATED","WAITING_PAYMENT","PAID","CANCELED"]},"totalAmount":{"type":"number"},"items":{"type":"array","items":{"$ref":"#/components/schemas/OrderItemResponseDTO"}}}},"ProductRequestDTO":{"required":["categoryId","name","price","stockQuantity"],"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"stockQuantity":{"type":"integer","format":"int32"},"categoryId":{"type":"string","format":"uuid"}}},"ProductImageDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"url":{"type":"string"},"mainImage":{"type":"boolean"}}},"ProductResponseDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"stockQuantity":{"type":"integer","format":"int32"},"active":{"type":"boolean"},"categoryName":{"type":"string"},"images":{"type":"array","items":{"$ref":"#/components/schemas/ProductImageDTO"}}}},"PagSeguroWebhookDTO":{"type":"object","properties":{"id":{"type":"string"},"status":{"type":"string"},"reference_id":{"type":"string"}}},"UserCreateDTO":{"required":["cpf","email","name","password","telefone"],"type":"object","properties":{"name":{"maxLength":100,"minLength":3,"type":"string"},"cpf":{"pattern":"\\d{11}","type":"string"},"telefone":{"maxLength":2147483647,"minLength":8,"type":"string"},"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":6,"type":"string"}}},"CheckoutRequestDTO":{"required":["orderId","paymentMethod"],"type":"object","properties":{"paymentMethod":{"type":"string","enum":["PIX","CREDIT_CARD","DEBIT_CARD"]},"orderId":{"type":"string","format":"uuid"},"cardToken":{"type":"string"},"installments":{"minimum":1,"type":"integer","format":"int32"}}},"CheckoutResponseDTO":{"type":"object","properties":{"paymentId":{"type":"string","format":"uuid"},"status":{"type":"string","enum":["PENDING","PAID","FAILED","CANCELED"]},"message":{"type":"string"},"pixQrCode":{"type":"string"},"pixCopyPaste":{"type":"string"}}},"AddItemRequestDTO":{"required":["productId","quantity"],"type":"object","properties":{"productId":{"type":"string","format":"uuid"},"quantity":{"minimum":1,"type":"integer","format":"int32"}}},"LoginRequestDTO":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"LoginResponseDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["ADMIN","CLIENT"]},"token":{"type":"string"}}},"AdminLoginRequestDTO":{"required":["adminCode","email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"adminCode":{"type":"string"}}},"ProductImageRequestDTO":{"required":["url"],"type":"object","properties":{"url":{"type":"string"},"mainImage":{"type":"boolean"}}}}}}