# BackLegiaoTricolor/src/main/java/com/legiaotricolor/BackLegiaoTricolor/controller/CheckoutController.java
package com.legiaotricolor.BackLegiaoTricolor.controller;

import com.legiaotricolor.BackLegiaoTricolor.domain.User;
import com.legiaotricolor.BackLegiaoTricolor.dto.CheckoutRequestDTO;
import com.legiaotricolor.BackLegiaoTricolor.dto.CheckoutResponseDTO;
import com.legiaotricolor.BackLegiaoTricolor.exception.BusinessException;
import com.legiaotricolor.BackLegiaoTricolor.service.PaymentService;
import jakarta.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.annotation.AuthenticationPrincipal;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/checkout")
public class CheckoutController {

    private final PaymentService paymentService;

    public CheckoutController(PaymentService paymentService) {
        this.paymentService = paymentService;
    }

    @PostMapping
    public ResponseEntity<CheckoutResponseDTO> checkout(
            @AuthenticationPrincipal User user,
            @Valid @RequestBody CheckoutRequestDTO dto
    ) throws BusinessException {

        return ResponseEntity.ok(paymentService.checkout(user, dto));
    }
}