# BackLegiaoTricolor/src/main/java/com/legiaotricolor/BackLegiaoTricolor/dto/AdminLoginRequestDTO.java
package com.legiaotricolor.BackLegiaoTricolor.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record AdminLoginRequestDTO(

        @Email
        @NotBlank
        String email,

        @NotBlank
        String password,

        @NotBlank
        String adminCode
) {}